fix: remove promise from umnt and move error enum to mnt
