+++ src/mount/umnt.rs
@@ -7,12 +7,6 @@ use std::path;
 
 use async_trait::async_trait;
 
-/// Defines callback to pass [`Umnt::umnt`] result into.
-#[async_trait]
-pub trait Promise {
-    async fn keep();
-}
-
 #[async_trait]
 pub trait Umnt {
     /// Removes the mount list entry for the directory that was
@@ -23,5 +17,5 @@ pub trait Umnt {
     ///
     /// AUTH_UNIX authentication or better is required.
     /// There are no MOUNT protocol errors which can be returned from this procedure.
-    async fn umnt(&self, dirpath: path::PathBuf, promise: impl Promise);
+    async fn umnt(&self, dirpath: path::PathBuf);
 }
