+++ src/mount/umntall.rs
@@ -5,12 +5,6 @@
 
 use async_trait::async_trait;
 
-/// Defines callback to pass [`Umntall::umntall`] result into.
-#[async_trait]
-pub trait Promise {
-    async fn keep();
-}
-
 #[async_trait]
 pub trait Umntall {
     /// Removes all of the mount entries for this client previously
@@ -18,5 +12,5 @@ pub trait Umntall {
     ///
     /// AUTH_UNIX authentication or better is required.
     /// There are no MOUNT protocol errors which can be returned from this procedure.
-    async fn umntall(&self, promise: impl Promise);
+    async fn umntall(&self);
 }
